Factor out common syncronous executor code
This paves the way for python35 asyncio executor implementation.
